Move season to a separate module.
